requirements.txt
adagio==0.2.4
aiohttp==1.0.5
alembic==1.8.1
antlr4-python3-runtime==4.10
anyio==3.6.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-generator==1.10
async-timeout==4.0.2
atpublic==3.0.1
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
beakerx==2.3.13
beakerx-base==2.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bottle==0.12.23
Brotli==1.0.9
bs4==0.0.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.6.15
certipy==0.1.3
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cryptography==37.0.4
cycler==0.11.0
da-vinci==0.3.0
dash==2.6.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2022.8.0
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
discord.py==0.16.12
Django==4.1
django-bootstrap-icons==0.8.1
django-bootstrap-v5==1.0.11
django-extensions==3.2.0
django-markup==1.6
django-steamauth==1.1.2.3
docutils==0.19
duckdb==0.4.0
dunamai==1.12.0
easy-thumbnails==2.8.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
flake8==5.0.4
Flask==2.2.2
Flask-Compress==1.12
fonttools==4.34.4
fs==2.4.16
fsspec==2022.7.1
fugue==0.7.1
fugue-sql-antlr==0.1.0
greenlet==1.1.2
ibis-framework==3.1.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.8.0
ipdb==0.13.9
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipython-sql==0.4.1
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.9
jsonschema==4.10.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-contrib-core==0.4.0
jupyter-core==4.11.1
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.5.0
jupyter-server==1.18.1
jupyter-telemetry==0.1.0
jupyterhub==2.3.1
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kiwisolver==1.4.4
locket==1.0.0
lxml==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
multidict==6.0.2
multipledispatch==0.6.0
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.2
oauthlib==3.2.0
packaging==21.3
pamela==1.0.0
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
parsy==1.4.0
partd==1.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
plotly==5.10.0
poetry-dynamic-versioning==0.17.1
prettytable==0.7.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
pyarrow==9.0.0
pycodestyle==2.9.1
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.9.2
pyflakes==2.5.0
Pygments==2.13.0
pyjsparser==2.7.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
pyspark==3.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.4
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qpd==0.3.3
qtconsole==5.3.1
QtPy==2.2.0
regex==2022.7.25
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scipy==1.9.0
Send2Trash==1.8.0
shortuuid==1.0.9
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
sqlparse==0.4.2
stack-data==0.3.0
steam==1.3.0
tabulate==0.8.10
tenacity==8.0.1
terminado==0.15.0
testpath==0.6.0
tinycss2==1.1.1
toml==0.10.2
tomlkit==0.11.4
toolz==0.12.0
tornado==6.2
traitlets==5.3.0
triad==0.6.7
typing_extensions==4.3.0
urllib3==1.26.11
vdf==3.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==3.4
Werkzeug==2.2.2
widgetsnbextension==3.6.1
yarl==1.8.1
zipp==3.8.1
